+++ interceptor/metrics/otelmetrics.go
@@ -9,7 +9,7 @@ import (
 	api "go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/sdk/metric"
 	"go.opentelemetry.io/otel/sdk/resource"
-	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
+	semconv "go.opentelemetry.io/otel/semconv/v1.37.0"
 
 	"github.com/kedacore/http-add-on/pkg/build"
 )
@@ -31,8 +31,8 @@ func NewOtelMetrics(options ...metric.Option) *OtelMetrics {
 	if options == nil {
 		res := resource.NewWithAttributes(
 			semconv.SchemaURL,
-			semconv.ServiceNameKey.String("interceptor-proxy"),
-			semconv.ServiceVersionKey.String(build.Version()),
+			semconv.ServiceName("interceptor-proxy"),
+			semconv.ServiceVersion(build.Version()),
 		)
 
 		options = []metric.Option{