+++ CHANGELOG.md
@@ -50,6 +50,7 @@ This changelog keeps track of work items that have been completed and are ready 
 - **CI**: Replace stale bot with official GitHub Actions stale action
 - **CI**: Use GitHub-hosted ARM64 runners for e2e tests ([#1388](https://github.com/kedacore/http-add-on/issues/1388))
 - **DevContainer**: Fix devcontainer build by updating deprecated Go tools ([#1347](https://github.com/kedacore/http-add-on/issues/1347))
+- **General** update OpenTelemetry semconv version to v1.37.0
 
 ## v0.11.1